[GNIP #3485] - GNIP: Add support for vector temporal series to GeoNode (CSV/SHP). #3485

+++ geonode/upload/upload.py
@@ -39,8 +39,7 @@ from geonode.layers.models import Layer
 from geonode import GeoNodeException
 from geonode.people.utils import get_default_user
 from geonode.upload.models import Upload
-from geonode.upload import signals
-from geonode.upload.utils import create_geoserver_db_featurestore
+from geonode.upload import signals, utils
 from geonode.geoserver.helpers import gs_catalog, gs_uploader, ogc_server_settings
 from geonode.geoserver.helpers import mosaic_delete_first_granule, set_time_dimension, set_time_info
 
@@ -60,6 +59,10 @@ import uuid
 import zipfile
 
 logger = logging.getLogger(__name__)
+
+
+def _log(msg, *args):
+    logger.info(msg, *args)
 
 
 class UploadException(Exception):
@@ -124,6 +127,9 @@ class UploaderSession(object):
     # track the most recently completed upload step
     completed_step = None
 
+    # track the most recently completed upload step
+    error_msg = None
+
     # the upload type - see the _pages dict in views
     upload_type = None
 
@@ -180,7 +186,6 @@ def upload(
         user = get_default_user()
     if isinstance(user, basestring):
         user = get_user_model().objects.get(username=user)
-
     import_session = save_step(
         user,
         name,
@@ -195,7 +200,6 @@ def upload(
         time_presentation_res=time_presentation_res,
         time_presentation_default_value=time_presentation_default_value,
         time_presentation_reference_value=time_presentation_reference_value)
-
     upload_session = UploaderSession(
         base_file=base_file,
         name=name,
@@ -209,7 +213,6 @@ def upload(
         mosaic_time_regex=mosaic_time_regex,
         mosaic_time_value=mosaic_time_value
     )
-
     time_step(upload_session,
               time_attribute, time_transform_type,
               presentation_strategy, precision_value, precision_step,
@@ -217,13 +220,9 @@ def upload(
               end_time_transform_type=end_time_transform_type,
               time_format=None, srs=None, use_big_date=use_big_date)
 
-    run_import(upload_session, async=False)
+    utils.run_import(upload_session, async=False)
 
     final_step(upload_session, user)
-
-
-def _log(msg, *args):
-    logger.info(msg, *args)
 
 
 def save_step(
@@ -260,41 +259,41 @@ def save_step(
     # Check if the store exists in geoserver
     try:
         store = gs_catalog.get_store(name)
+
+        if store:
+            # If we get a store, we do the following:
+            resources = store.get_resources()
+            # Is it empty?
+            if len(resources) == 0:
+                # What should we do about that empty store?
+                if overwrite:
+                    # We can just delete it and recreate it later.
+                    store.delete()
+                else:
+                    msg = (
+                        'The layer exists and the overwrite parameter is %s' %
+                        overwrite)
+                    raise GeoNodeException(msg)
+            else:
+
+                # If our resource is already configured in the store it
+                # needs to have the right resource type
+                for resource in resources:
+                    if resource.name == name:
+
+                        assert overwrite, "Name already in use and overwrite is False"
+
+                        existing_type = resource.resource_type
+                        if existing_type != the_layer_type:
+                            msg = (
+                                'Type of uploaded file %s (%s) does not match type '
+                                'of existing resource type %s' %
+                                (name, the_layer_type, existing_type))
+                            _log(msg)
+                            raise GeoNodeException(msg)
     except geoserver.catalog.FailedRequestError as e:
         # There is no store, ergo the road is clear
         pass
-    else:
-        # If we get a store, we do the following:
-        resources = store.get_resources()
-        # Is it empty?
-        if len(resources) == 0:
-            # What should we do about that empty store?
-            if overwrite:
-                # We can just delete it and recreate it later.
-                store.delete()
-            else:
-                msg = (
-                    'The layer exists and the overwrite parameter is %s' %
-                    overwrite)
-                raise GeoNodeException(msg)
-        else:
-
-            # If our resource is already configured in the store it
-            # needs to have the right resource type
-
-            for resource in resources:
-                if resource.name == name:
-
-                    assert overwrite, "Name already in use and overwrite is False"
-
-                    existing_type = resource.resource_type
-                    if existing_type != the_layer_type:
-                        msg = (
-                            'Type of uploaded file %s (%s) does not match type '
-                            'of existing resource type %s' %
-                            (name, the_layer_type, existing_type))
-                        _log(msg)
-                        raise GeoNodeException(msg)
 
     if the_layer_type not in (
             FeatureType.resource_type,
@@ -377,9 +376,10 @@ def save_step(
 
         if not import_session.tasks:
             error_msg = 'No valid upload files could be found'
-        elif import_session.tasks[0].state == 'NO_FORMAT':
+        elif import_session.tasks[0].state == 'NO_FORMAT' \
+                or import_session.tasks[0].state == 'BAD_FORMAT':
             error_msg = 'There may be a problem with the data provided - ' \
-                        'we could not identify it'
+                        'we could not identify its format'
 
         if len(import_session.tasks) > 1:
             error_msg = "Only a single upload is supported at the moment"
@@ -412,63 +412,6 @@ def save_step(
         _log("Finished upload of [%s] to GeoServer without errors.", name)
 
     return import_session
-
-
-def run_import(upload_session, async):
-    """Run the import, possibly asynchronously.
-
-    Returns the target datastore.
-    """
-    import_session = upload_session.import_session
-    import_session = gs_uploader.get_session(import_session.id)
-    task = import_session.tasks[0]
-    import_execution_requested = False
-    if import_session.state == 'INCOMPLETE':
-        if task.state != 'ERROR':
-            raise Exception('unknown item state: %s' % task.state)
-    elif import_session.state == 'PENDING' and task.target.store_type == 'coverageStore':
-        if task.state == 'READY':
-            import_session.commit(async)
-            import_execution_requested = True
-
-    # if a target datastore is configured, ensure the datastore exists
-    # in geoserver and set the uploader target appropriately
-
-    if ogc_server_settings.GEOGIG_ENABLED and upload_session.geogig is True \
-            and task.target.store_type != 'coverageStore':
-        target = create_geoserver_db_featurestore(
-            store_type='geogig',
-            store_name=upload_session.geogig_store,
-            author_name=upload_session.user.username,
-            author_email=upload_session.user.email)
-        _log(
-            'setting target datastore %s %s',
-            target.name,
-            target.workspace.name)
-        task.set_target(target.name, target.workspace.name)
-
-    elif ogc_server_settings.datastore_db and task.target.store_type != 'coverageStore':
-        target = create_geoserver_db_featurestore()
-        _log(
-            'setting target datastore %s %s',
-            target.name,
-            target.workspace.name)
-        task.set_target(target.name, target.workspace.name)
-    else:
-        target = task.target
-
-    if upload_session.update_mode:
-        _log('setting updateMode to %s', upload_session.update_mode)
-        task.set_update_mode(upload_session.update_mode)
-
-    _log('running import session')
-    # run async if using a database
-    if not import_execution_requested:
-        import_session.commit(async)
-
-    # @todo check status of import session - it may fail, but due to protocol,
-    # this will not be reported during the commit
-    return target
 
 
 def time_step(upload_session, time_attribute, time_transform_type,
@@ -515,7 +458,6 @@ def time_step(upload_session, time_attribute, time_transform_type,
 
     if time_attribute:
         if time_transform_type:
-
             transforms.append(
                 build_time_transform(
                     time_attribute,
@@ -524,7 +466,6 @@ def time_step(upload_session, time_attribute, time_transform_type,
             )
 
         if end_time_attribute and end_time_transform_type:
-
             transforms.append(
                 build_time_transform(
                     end_time_attribute,
@@ -534,7 +475,6 @@ def time_step(upload_session, time_attribute, time_transform_type,
 
         # this must go after the remapping transform to ensure the
         # type change is applied
-
         if use_big_date:
             transforms.append(build_att_remap_transform(time_attribute))
             if end_time_attribute:
@@ -568,31 +508,45 @@ def time_step(upload_session, time_attribute, time_transform_type,
         upload_session.import_session.tasks[0].add_transforms(transforms)
         try:
             upload_session.time_transforms = transforms
+            upload_session.time = True
         except BadRequest as br:
             raise UploadException.from_exc('Error configuring time:', br)
         upload_session.import_session.tasks[0].save_transforms()
+    else:
+        upload_session.time = False
 
 
 def csv_step(upload_session, lat_field, lng_field):
     import_session = upload_session.import_session
     task = import_session.tasks[0]
+
     transform = {'type': 'AttributesToPointGeometryTransform',
                  'latField': lat_field,
                  'lngField': lng_field,
                  }
-    task.layer.set_srs('EPSG:4326')
     task.remove_transforms([transform], by_field='type', save=False)
     task.add_transforms([transform], save=False)
     task.save_transforms()
+    import_session = import_session.reload()
+    upload_session.import_session = import_session
 
 
-def srs_step(upload_session, srs):
-    layer = upload_session.import_session.tasks[0].layer
-    srs = srs.strip().upper()
-    if not srs.startswith("EPSG:"):
-        srs = "EPSG:%s" % srs
-    logger.info('Setting SRS to %s', srs)
-    layer.set_srs(srs)
+def srs_step(upload_session, source, target):
+    import_session = upload_session.import_session
+    task = import_session.tasks[0]
+    if source:
+        logger.info('Setting SRS to %s', source)
+        task.set_srs(source)
+
+    transform = {'type': 'ReprojectTransform',
+                 'source': source,
+                 'target': target,
+                 }
+    task.remove_transforms([transform], by_field='type', save=False)
+    task.add_transforms([transform], save=False)
+    task.save_transforms()
+    import_session = import_session.reload()
+    upload_session.import_session = import_session
 
 
 def final_step(upload_session, user):
@@ -624,8 +578,11 @@ def final_step(upload_session, user):
     if import_session.state == 'INCOMPLETE':
         if task.state != 'ERROR':
             raise Exception('unknown item state: %s' % task.state)
+    elif import_session.state == 'READY':
+        import_session.commit()
     elif import_session.state == 'PENDING':
-        if task.state == 'READY' and task.data.format != 'Shapefile':
+        if task.state == 'READY':
+            # if not task.data.format or task.data.format != 'Shapefile':
             import_session.commit()
 
     if not publishing:
@@ -641,9 +598,25 @@ def final_step(upload_session, user):
         base_file = upload_session.base_file
         sld_file = base_file[0].sld_files[0]
 
-        f = open(sld_file, 'r')
-        sld = f.read()
-        f.close()
+        f = None
+        if os.path.isfile(sld_file):
+            try:
+                f = open(sld_file, 'r')
+            except:
+                pass
+        elif upload_session.tempdir and os.path.exists(upload_session.tempdir):
+            tempdir = upload_session.tempdir
+            if os.path.isfile(os.path.join(tempdir, sld_file)):
+                try:
+                    f = open(os.path.join(tempdir, sld_file), 'r')
+                except:
+                    pass
+
+        if f:
+            sld = f.read()
+            f.close()
+        else:
+            sld = get_sld_for(cat, publishing)
     else:
         sld = get_sld_for(cat, publishing)
 
@@ -755,6 +728,8 @@ def final_step(upload_session, user):
                     'There was an error updating the mosaic temporal extent: ' +
                     str(e))
     else:
+        _has_time = (True if upload_session.time and upload_session.time_info and
+                     upload_session.time_transforms else False)
         saved_layer, created = Layer.objects.get_or_create(
             name=task.layer.name,
             defaults=dict(store=target.name,
@@ -764,7 +739,8 @@ def final_step(upload_session, user):
                           title=title,
                           uuid=layer_uuid,
                           abstract=abstract or '',
-                          owner=user,)
+                          owner=user,),
+            has_time=_has_time
         )
 
     # Should we throw a clearer error here?
